scraper/requirements.txt
playwright
firebase-admin
python-dotenv
pandas
openpyxl
